Remove Donations + Sponsorships
I'm not running the demo server anymore because it's too costly at the moment for me to host it all by myself. I believe this README suffices for understanding how this open-source repo works. If someone else wants to run a demo server, I'd be happy to link it in the README.